update testing module
