+++ mindee/error/__init__.py
@@ -1,1 +1,6 @@
 from mindee.error.mindee_error import MindeeClientError, MindeeError
+from mindee.error.mindee_http_error import (
+    MindeeHTTPClientError,
+    MindeeHTTPError,
+    MindeeHTTPServerError,
+)
