+++ mindee/error/mindee_error.py
@@ -6,5 +6,9 @@ class MindeeClientError(MindeeError):
     """An exception relating to document parsing."""
 
 
+class MindeeApiError(MindeeError):
+    """An exception relating to settings of the MindeeClient."""
+
+
 class MindeeSourceError(MindeeError):
     """An exception relating to document loading."""
